[PaymentHandler] Support lack of payment handler icon in minimal UX
This patch adds support for a lack of an icon in a web-based payment
handler when the PaymentHandlerMinimalHeaderUX feature is enabled. This
is implemented by only adding the icon image to the header if it is
present and non-empty, and otherwise adding a padding column in its
place.

This is tested by the existing PaymentHandlerHeaderViewUITest
HeaderWithoutIcon test by removing the early-exit when minimal UX is
enabled.

Bug: 1414084
Change-Id: I34cf967e6b42e610e82a6230973014f63b1ede4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4225056
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102810}