Added noise cancellation toggle for QsRevamp
In this CL, the noise cancellation toggle button is added in the audio
subpage. The row has an icon, a text label, and a toggle button. The
entire row should be focusable and togglable. A separator is added if
the input slider is not the last one in the list. Existing unit tests
are modified for the revamped view.

Screenshots:
Toggled-on v.s. toggled-off:
https://screenshot.googleplex.com/4nNZMp6onGzdf2P

Separator when there's a following slider:
https://screenshot.googleplex.com/53gecDiDo5jRkch

Focus on the row:
https://screenshot.googleplex.com/4PqNbv32wxxYS7L

Bug: b/266996145
Change-Id: I4a8c4baca13ee7c8fac8dd2bc6369c70a72f4256
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4222066
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102891}