Fix PWA/SWA bounds when recent window is minimized
When the most recent window is minimized, the bound set by
GetLastActiveWindowState can be empty.  This causes windows to appear
tiny and in the top left corner of the screen.  Work around this
behavior by triggering the no-windows codepath if the bounds
are empty.

Bug: b/266865594
Change-Id: Ieb6492fd6b61de967e8696822320dbdad8d4b1d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220297
Auto-Submit: Sean Kau <skau@chromium.org>
Commit-Queue: Sean Kau <skau@chromium.org>
Reviewed-by: Dominic Schulz <dominicschulz@google.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102938}