[PaymentHandler] Set accessibility text for payment handler icon
This patch updates the accessibility text of the payment handler icon to
be consistent with other icons in Chrome, i.e. simply read out that it
is an icon. Previously, the text was set to be the payment handler
title, which isn't very accurate as the title is not actually shown.

Screenshot provided for translation:
https://storage.cloud.google.com/chromium-translation-screenshots/a4914f944acc47c8932fe1646a177f216f3b28cc

Bug: 1414090
Change-Id: I04e1e14d0fe04639774e6c811d94d15c8328b715
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4228692
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102952}