[Thumbnail Cache] Disable approximation cache with flag
This is the first in a series of flag guarded refactorings to the
TabContentManager and ThumbnailCache.

This adds the flag and disables the approximation cache. These
approximations are 1/4 scale (1/16th size) bitmaps of tab thumbnails.
A modern high-end device each represents ~1 MB so the removal of
this effectively unused cache represents an estimated savings of 8 MB.

Bug: 1402843
Change-Id: Ibd852dc6c9ae52b21547a4fe8f4ac923af0a8fa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220431
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102970}