Align serialized PaintOp to maximum alignment of data
PaintOpWriter/PaintOpReader::AlignMemory() can be called with 16
from image_transform_cache_entry.cc, which requires the serialized
memory also to be aligned to 16 bytes, otherwise the starting
address of the memory can change the number of padding bytes used
to align to 16 bytes.

Now PaintOpWriter/PaintOpReader require the memory to be aligned to
16 bytes if enable_security_constraints is false, otherwise 4, and
their AlignMemory() checks the alignment can't exceed the alignment.
Changed a possible source of the memory WebGPUImplementation to
allocate memory aligned to 16 bytes.

Removed AlignMemory(8) in PaintOpWriter::WriteSize() and
PaintOpReader::ReadSize() because the alignment should never exceed
alignment 4 which is used when enable_security_constraints is true.
We have been always using alignment 4 for uint64_t, so using alignment
4 for size_t should be good, too.

PaintOpWriter/PaintOpReader are also used to read/write specific data
(e.g. PaintFilter in paint_filter_mojom_traits.cc) instead of a whole
PaintOp. In the case, now avoid the unnecessary op header. The
alignment requirement is still the same as above.

Bug: 1300188, 1336843
Change-Id: Ie0fff6266b7313ba7e074eda55bc1e0163eb5b74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4205862
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102975}