[FilesBulkPinning] Remove local source, stop calculating when cancelled
Setting the query to local source seems to miss a lot of files which are
actually available offline. From the query.proto (see b/268264938) the
availabe_offline bool actually constrains the query to the local cache
anyway, so only use this field.

Furthermore if a user closes the storage management page, the callback
WeakPtr gets invalidated but the calculation still continues. Use the
callbacks `IsCancelled` method as a proxy to abort the cancellation
early in that case.

Fixed: b:268264938
Test: browser_tests --gtest_filter=*GetTotalPinnedSize*
Change-Id: I4436b61ebc3f49df710c915937734da50f817bad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4230933
Reviewed-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102992}