Make SandboxedZipAnalyzer owned by FileAnalyzer
SandboxedZipAnalyzer is currently ref-counted, and creates a
reference to itself during analysis in the utility process. This
means that this object is not destroyed until the utility process
completes, which can take a long time in exceptional circumstances. To
ensure timely cleanup when enforcing a timeout on download inspection,
this CL makes the SandboxedZipAnalyzer owned by the FileAnalyzer. When
it is destroyed, the mojo::Remote to the FileUtilService will be
destroyed, stopping all work in the utility process and cleaning up the
temporary files.

Future CLs will make analogous changes to the other sandboxed analyzers

Bug: 1373679
Change-Id: I3ec5b978877e5e37f97c7d4dc034d04ee1883eab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4232867
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103066}