[Rounded Display] Adds RoundedDisplay Gutter class.
It is responsible to manage and create a single texture that
represents a single or two rounded corners of the display depending on
the gutter type.

These gutters will be managed by RoundedDisplayProvider and
RoundedDisplayHost will request these gutters to create textures as it
renders it to display in the follow-up CLs.

Bug: 1262970
Test: Covered by unittests.
Change-Id: I8ff435f552ed5ec1c6af13fe8f854780782ca381
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3894878
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103121}