[balance-text] Replace `WhiteSpace()` in `IsTabHTMLSpanElement`
This patch replaces `WhiteSpace() == Pre` in
`IsTabHTMLSpanElement` with more semantical function
`IsCollapsibleWhiteSpace` to prepare for refactoring the
`white-space` property to a shorthand.

This code was added in crrev.com/c/697044 to check if a TAB
character is preserved. As part of the efforts to change the
`white-space` property to a shorthand, this change eliminates
the direct access to the `WhiteSpace()` property.

This change will apply the fix in crrev.com/c/697044 not only
to `pre` but also to `pre-wrap` and `break-spaces` where TAB
characters are also preserved.

Bug: 1251079
Change-Id: I26c94f93800471803a285383ca6ca520740a12e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4233237
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103125}