[balance-text] Replace `WhiteSpace()` in `LayoutSVGInlineText`
This patch replaces `WhiteSpace()` with a semantics function
`CollapseWhiteSpace()` to prepare for refactoring the
`white-space` property to a shorthand.

This logic was introduced in 2010[1]. This change may cause
more calls to `ForceSetText` such as when `normal` is changed
to `pre-wrap` or `pre-line`, but it should be more correct.

[1] https://chromium.googlesource.com/chromium/src/+/8a3844a5c71c228eba5a014917f032324af30f69%5E%21/third_party/WebKit/WebCore/rendering/svg/RenderSVGInlineText.cpp

Bug: 1251079
Change-Id: I1da6e6aae2b5324db1b4cf3d1b04c4c94b3ca96e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4235799
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103133}