[copy_or_move] expand comment re automatic conflict resolves
Once the conflict_resolve_ member is set (it is not empty), the UI
is never asked to resolve a file name conflict again.

The way to resolve is known and stored in conflict_resolve_ and is
by definition, an automatic "Apply-to-all". Expand code comment to
mention these details.

Bug: b:255264604
Change-Id: I8a5c00365c51b7733ce34cadfaf0050ec1454338
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4234681
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103138}