Store calculated service worker script checksums to the database
This CL adds the mechanism to save the chekcsum hash strings of service
worker scripts, which are calculated while the service worker update
check process. After this CL, if the user has the service worker script
info in the database and its |sha256_checksum| is empty, the update
checker returns the sha256 checksums for both main and sub scripts,
regardless of the update check result.

Once the update check finished, the browser tries to update the script
resource list and store it to the database with the newly calculated
checksums.

We don't update the service worker version since the only updated thing
is the |sha256_checksum| record in DB and that is no user visible
change. However, this decision introduced more code changes, so using
`StoreRegistration`, or adding the option not to create the new version
to `StoreRegistration` may be another good option, at least in terms of
the code complexitiy. Feedback is welcomed.

Bug: 1371756
Change-Id: Ie4cf2033c0f06407e41e7d35fae048ce9741877c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210875
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103144}