[PolicyLogs] First Batch of Logs for chrome://policy/logs
This CL adds the first set of logs related to CBCM enrollment. Existing
logs using base/logging are now using the xLOG_POLICY macro
which will use base/logging and also add to the list of logs that are to
be displayed on chrome://policy/logs if the feature is enabled.


Bug: b/266480750, b/251799398
Change-Id: I82c53fa516c598f1aff1b6a18308d13674de5997
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4221504
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103150}