Remove DCHECK in RelayoutIgnoringChildScrollbarChanges().
We used to DCHECK that we didn't have layout_info_for_devtools_ set when
re-laying out after a scrollbar change. But this situation seems
harmless, and the DCHECK failure could be triggered quite easily. All we
needed was a flex item with an auto scrollbar, and toggle the need for
the scrollbar, while devtools was active.

NGFlexLayoutAlgorithm::NGFlexLayoutAlgorithm() will set up
layout_info_for_devtools_ if needed. If we end up in
RelayoutIgnoringChildScrollbarChanges(), we'll invoke another
NGFlexLayoutAlgorithm, which will set up and populate
layout_info_for_devtools_ again, transfer it to the fragment builder and
store it in the layout result. The output from the original flex layout
algorithm instance will be discarded and have no effect, on the other
hand. This situation doesn't seem problematic at all to me.

Also remove an unused DevtoolsFlexInfo* parameter to
NGFlexLayoutAlgorithm(), which in a way did suggest that it was bad to
just drop the original layout_info_for_devtools_ on the floor.

Bug: 1395114
Change-Id: I832ea0c59e0ad0fd09a7837f7ff798c03d32f304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4231753
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103172}