[Password Manager] Release PasswordsPrivateDelegate when not needed
This CLs allows to control lifetime of PasswordsPrivateDelegate
depending on Password Manager redesign feature state. When the new UI is
enabled PasswordsPrivateDelegate will live as long as somebody is
holding a scoped_refptr. This is achieved by creating a proxy object
between factory and delegate. This proxy is bound to the browser and
holds either weak_ptr or scoped_ptr depending on the feature state.

Bug: 1412348
Change-Id: Ic0b5c84026672cb48def8d0dc1bdd3d302e00d84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4205116
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103175}