webview/metrics: Fix non embedded metrics returning status
In theory this should help us with the metrics loss
we saw last time we tried to move metrics
reporting over to the non embedded process.

Only doing this for the non embedded
because we'd need to turn the embedded metrics reporting
into a blocking call which would introduce more risk
than we want.

In practice, this means we can do the following
experiments. We can roll out moving metric reporting to
a background thread, and we can then experiment with
the non embedded process within that.

Test: Tested this manually by enabling both
Test: flags and ran the webview_instrumentation_test_apk.
Bug: 1264425
Change-Id: Id7d764ca2066ad6a06a0982c0f7ab9266cfea6ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4221614
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Auto-Submit: Rupert Wiser <bewise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103192}