crostini: Wire up VmStopping signal
Instead of relying in Chrome to tell CrostiniManager when a VM is
stopping (so we don't know when it's vmc or otherwise triggered, let
Concierge tell us as the authority. We still note it on StopVm calls as
that lets us know slightly (1 dbus call) sooner.

Bug: b:245410022
Test: CQ
Change-Id: I6709efb8f12d2c2b3aab1cc835e90f6ff45004c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4223851
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: David Munro <davidmunro@google.com>
Cr-Commit-Position: refs/heads/main@{#1103214}