[Autofill] Prompt before removing a local credit card.
This CL adds a confirmation dialog to confirm a user's intent to remove
a locally stored credit card in settings. This aligns the experience for
removing local credit cards with that of addresses.

The CL also changes the word 'Remove' to 'Delete', in line with the
same change made to addresses in https://crrev.com/a5094829015

It was adapted from initial work done by schwering@ in
https://chromium-review.googlesource.com/c/chromium/src/+/3501833

Bug: 1301287
Change-Id: I294ce7faaa09265cefeff6de79d685385321cba9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219354
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103276}