[iOS] Fix GoogleServicesSettingsTestCase/testOpeningServices
The test was searching for 3 cells in the following ordier:
 * IDS_IOS_GOOGLE_SERVICES_SETTINGS_IMPROVE_SEARCH_SUGGESTIONS_TEXT
 * IDS_IOS_GOOGLE_SERVICES_SETTINGS_IMPROVE_CHROME_TEXT
 * IDS_IOS_GOOGLE_SERVICES_SETTINGS_BETTER_SEARCH_AND_BROWSING_TEXT

The problem is "search suggestion" is the last cell in the view, and
"improve chrome" is the second cell.
After finding "search suggestion", the scroll view is at the bottom.
And "improve chrome" is hidden.

The fix is to search for cells, the right order:
 * IDS_IOS_GOOGLE_SERVICES_SETTINGS_IMPROVE_CHROME_TEXT
 * IDS_IOS_GOOGLE_SERVICES_SETTINGS_BETTER_SEARCH_AND_BROWSING_TEXT
 * IDS_IOS_GOOGLE_SERVICES_SETTINGS_IMPROVE_SEARCH_SUGGESTIONS_TEXT

Also, -[GoogleServicesSettingsTestCase assertNonPersonalizedServices]
is inlined in this patch since it is called only from one place.

Fixed: 1414307
Change-Id: Idf287f422f368957647091cb023c7fdb87c7b26a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4233841
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Elmehdi Rahmaoui <erahmaoui@google.com>
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Elmehdi Rahmaoui <erahmaoui@google.com>
Cr-Commit-Position: refs/heads/main@{#1103285}