media: Ensure GpuResources has last shared image ref
On context loss or stub destruction, the WillDestroyStub callback will
cause GpuResources to destroy all shared image representations that it
holds. Later DefaultTexture2DWrapper destructor will drop its shared
image representation which will cause the shared image to be destroyed.
There are two issues with this: 1) it happens on the decoder sequence
instead of gpu sequence, and 2) in case of context loss the information
isn't propagated to the shared image.

This CL ensures that the last ref is dropped by GpuResources by keeping
its shared image representation alive until GpuResources destructor. In
addition we have to be careful in propagating the correct context lost
state to avoid a memory leak.

Bug: 1412751
Change-Id: I1cfad933790bf1c3055c17e9f19deace5d16187a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4234826
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103349}